Added buttons to LFG posts

+++ cogs/group.py
@@ -46,7 +46,11 @@ class Group(commands.Cog):
         if is_embed and ctx.channel.permissions_for(ctx.guild.me).embed_links:
             embed = ctx.bot.raid.make_embed(message, ctx.bot.translations[lang], lang)
             out = await message.channel.send(content=msg)
-            await out.edit(content=None, embed=embed)
+            buttons = GroupButtons(out.id, ctx.bot, label_go=ctx.bot.translations[lang]['lfg']['button_want'],
+                                   label_help=ctx.bot.translations[lang]['lfg']['button_help'],
+                                   label_delete=ctx.bot.translations[lang]['lfg']['button_no_go'],
+                                   label_no_go=ctx.bot.translations[lang]['lfg']['button_delete'])
+            await out.edit(content=None, embed=embed, view=buttons)
         else:
             out = await message.channel.send(msg)
         ctx.bot.raid.set_id(out.id, message.id)
@@ -279,9 +283,9 @@ class Group(commands.Cog):
                 group_ch_id = group_ch.id
             ctx.bot.raid.set_group_space(group_id, group_role.id, group_ch_id)
         out = await ctx.channel.fetch_message(group_id)
-        await out.add_reaction('👌')
-        await out.add_reaction('❓')
-        await out.add_reaction('❌')
+        # await out.add_reaction('👌')
+        # await out.add_reaction('❓')
+        # await out.add_reaction('❌')
 
     def parse_date(self, time):
         try:
@@ -341,12 +345,26 @@ class Group(commands.Cog):
             dm = await ctx.message.create_thread(name='LFG', auto_archive_duration=60)
             await dm.add_user(ctx.message.author)
 
-            lfg_list = await self.lfglist(ctx, lang)
-            if not lfg_list:
-                if type(dm) == nextcord.Thread:
-                    await asyncio.sleep(10)
-                    await dm.delete()
-                return
+            lfg_list = ctx.bot.raid.c.execute(
+                'SELECT group_id, name, time, channel_name, server_name, timezone FROM raid WHERE owner=?',
+                (ctx.message.author.id,))
+            lfg_list = lfg_list.fetchall()
+
+            msg = translations['lfglist_head']
+            i = 1
+            for lfg in lfg_list:
+                msg = translations['lfglist'].format(msg, i, lfg[1], datetime.fromtimestamp(lfg[2]), lfg[5],
+                                                     lfg[3], lfg[4], ctx.bot.raid.hashids.encode(lfg[0]))
+                i = i + 1
+            if len(lfg_list) > 0:
+                await dm.send(msg)
+            else:
+                await dm.send(translations['lfglist_empty'])
+                if not lfg_list:
+                    if type(dm) == nextcord.Thread:
+                        await asyncio.sleep(10)
+                        await dm.delete()
+                    return
             number = await get_numerical_answer('lfg_choice', len(lfg_list) + 1)
         else:
             if ctx.author.dm_channel is None:
@@ -615,7 +633,12 @@ class Group(commands.Cog):
             if old_lfg is not None and owner is not None:
                 old_lfg = await old_lfg.fetch_message(group_id[0])
                 if owner == message.author.id:
-                    await ctx.bot.raid.edit(message, old_lfg, ctx.bot.translations[lang], lang, text)
+                    new_lfg = await ctx.bot.raid.edit(message, old_lfg, ctx.bot.translations[lang], lang, text)
+                    buttons = GroupButtons(new_lfg.id, ctx.bot, label_go=ctx.bot.translations[lang]['lfg']['button_want'],
+                                           label_help=ctx.bot.translations[lang]['lfg']['button_help'],
+                                           label_delete=ctx.bot.translations[lang]['lfg']['button_no_go'],
+                                           label_no_go=ctx.bot.translations[lang]['lfg']['button_delete'])
+                    await new_lfg.edit(view=buttons)
                 else:
                     await ctx.bot.check_ownership(message)
                     if ctx.channel.permissions_for(ctx.guild.me).manage_messages:
@@ -737,6 +760,97 @@ class RoleLFG(nextcord.ui.View):
         self.add_item(self.custom_button)
         self.add_item(self.default_button)
         self.value = None
+
+
+class WantButton(nextcord.ui.Button):
+    def __init__(self, label, style, custom_id, row=1):
+        super().__init__(style=style, label=label, row=row, custom_id=custom_id)
+
+    async def callback(self, interaction: nextcord.Interaction):
+        self.view.bot.raid.add_people(interaction.message.id, interaction.user)
+        lang = self.view.bot.guild_lang(interaction.message.guild.id)
+        await self.view.bot.raid.update_group_msg(interaction.message, self.view.bot.translations[lang], lang)
+        mode = self.view.bot.raid.get_cell('group_id', interaction.message.id, 'group_mode')
+        owner = self.view.bot.get_user(self.view.bot.raid.get_cell('group_id', interaction.message.id, 'owner'))
+        if mode == 'manual' and owner.id != interaction.user.id:
+            if interaction.user.nick is not None:
+                nick = interaction.user.nick
+            else:
+                nick = interaction.user.name
+            await interaction.response.send_message(content=self.view.bot.translations[lang]['lfg']['gotcha'].format(nick), ephemeral=True)
+            await self.view.bot.raid.upd_dm(owner, interaction.message.id, self.view.bot.translations[lang])
+
+
+class MaybeButton(nextcord.ui.Button):
+    def __init__(self, label, style, custom_id, row=1):
+        super().__init__(style=style, label=label, row=row, custom_id=custom_id)
+
+    async def callback(self, interaction: nextcord.Interaction):
+        self.view.bot.raid.add_mb_goers(interaction.message.id, interaction.user)
+        lang = self.view.bot.guild_lang(interaction.message.guild.id)
+        await self.view.bot.raid.update_group_msg(interaction.message, self.view.bot.translations[lang], lang)
+
+
+class NoGoButton(nextcord.ui.Button):
+    def __init__(self, label, style, custom_id, row=1):
+        super().__init__(style=style, label=label, row=row, custom_id=custom_id)
+
+    async def callback(self, interaction: nextcord.Interaction):
+        was_goer = self.view.bot.raid.is_goer(interaction.message, interaction.user)
+        is_mb_goer = self.view.bot.raid.is_mb_goer(interaction.message, interaction.user)
+        emoji = ''
+        if was_goer:
+            emoji = '👌'
+        elif is_mb_goer:
+            emoji = '❓'
+        self.view.bot.raid.rm_people(interaction.message.id, interaction.user, emoji)
+        lang = self.view.bot.guild_lang(interaction.message.guild.id)
+        await self.view.bot.raid.update_group_msg(interaction.message, self.view.bot.translations[lang], lang)
+
+
+class DeleteButton(nextcord.ui.Button):
+    def __init__(self, label, style, custom_id, row=1):
+        super().__init__(style=style, label=label, row=row, custom_id=custom_id)
+
+    async def callback(self, interaction: nextcord.Interaction):
+        owner = self.view.bot.get_user(self.view.bot.raid.get_cell('group_id', interaction.message.id, 'owner'))
+        message = interaction.message
+        if owner.id == interaction.user.id:
+            mode = self.view.bot.raid.get_cell('group_id', message.id, 'group_mode')
+            if mode == 'manual':
+                dm_id = self.view.bot.raid.get_cell('group_id', message.id, 'dm_message')
+                if owner.dm_channel is None:
+                    await owner.create_dm()
+                if dm_id != 0:
+                    dm_message = await owner.dm_channel.fetch_message(dm_id)
+                    await dm_message.delete()
+            if message.guild.me.guild_permissions.manage_roles:
+                role = message.guild.get_role(self.view.bot.raid.get_cell('group_id', message.id, 'group_role'))
+                if role is not None:
+                    await role.delete(reason='LFG deletion')
+            if message.guild.me.guild_permissions.manage_channels:
+                group_ch = message.guild.get_channel(self.view.bot.raid.get_cell('group_id', message.id, 'group_channel'))
+                if group_ch is not None:
+                    if group_ch.permissions_for(message.guild.me).manage_channels:
+                        await group_ch.delete(reason='LFG deletion')
+            self.view.bot.raid.del_entry(message.id)
+            await message.delete()
+
+
+class GroupButtons(nextcord.ui.View):
+
+    def __init__(self, group_id, bot, label_go='👌', label_help='❓', label_delete='❌', label_no_go='-'):
+        super().__init__(timeout=None)
+        self.bot = bot
+
+        self.want_button = WantButton(label_go, nextcord.ButtonStyle.green, '{}_go'.format(group_id))
+        self.maybe_button = MaybeButton(label_help, nextcord.ButtonStyle.gray, '{}_maybe'.format(group_id))
+        self.no_go_button = NoGoButton(label_no_go, nextcord.ButtonStyle.gray, '{}_no_go'.format(group_id))
+        self.delete_button = DeleteButton(label_delete, nextcord.ButtonStyle.red, '{}_delete'.format(group_id))
+        self.add_item(self.want_button)
+        self.add_item(self.maybe_button)
+        self.add_item(self.no_go_button)
+        self.add_item(self.delete_button)
 
 
 def setup(bot):
